# [mypyc] Add a special case for str_rprimitive in builder.builtin_len