--- a/mypyc/lib-rt/str_ops.c
+++ b/mypyc/lib-rt/str_ops.c
@@ -108,3 +108,7 @@ bool CPyStr_IsTrue(PyObject *obj) {
     Py_ssize_t length = PyUnicode_GET_LENGTH(obj);
     return length != 0;
 }
+
+Py_ssize_t CPyStr_Size_size_t(PyObject *str) {
+     return PyUnicode_GET_LENGTH(str);
+}