--- a/mypyc/primitives/str_ops.py
+++ b/mypyc/primitives/str_ops.py
@@ -139,6 +139,6 @@ str_check_if_true = custom_op(
 str_size_op = custom_op(
     arg_types=[str_rprimitive],
     return_type=c_pyssize_t_rprimitive,
-    c_function_name='PyUnicode_Size',
+    c_function_name='PyUnicode_GET_LENGTH',
     error_kind=ERR_NEVER)