--- a/mypyc/primitives/str_ops.py
+++ b/mypyc/primitives/str_ops.py
@@ -89,7 +89,7 @@ for i in range(len(str_split_types)):
 
 # str1 += str2
 #
-# PyUnicodeAppend makes an effort to reuse the LHS when the refcount
+# PyUnicode_Append makes an effort to reuse the LHS when the refcount
 # is 1. This is super dodgy but oh well, the interpreter does it.
 binary_op(name='+=',
           arg_types=[str_rprimitive, str_rprimitive],
@@ -116,7 +116,7 @@ method_op(
     name='replace',
     arg_types=[str_rprimitive, str_rprimitive, str_rprimitive],
     return_type=str_rprimitive,
-    c_function_name="PyUnicode_Replace",
+    c_function_name='PyUnicode_Replace',
     error_kind=ERR_MAGIC,
     extra_int_constants=[(-1, c_int_rprimitive)])
 
@@ -125,19 +125,19 @@ method_op(
     name='replace',
     arg_types=[str_rprimitive, str_rprimitive, str_rprimitive, int_rprimitive],
     return_type=str_rprimitive,
-    c_function_name="CPyStr_Replace",
+    c_function_name='CPyStr_Replace',
     error_kind=ERR_MAGIC)
 
 # check if a string is true (isn't an empty string)
 str_check_if_true = custom_op(
     arg_types=[str_rprimitive],
     return_type=bit_rprimitive,
-    c_function_name="CPyStr_IsTrue",
+    c_function_name='CPyStr_IsTrue',
     error_kind=ERR_NEVER,
 )
 
 str_ssize_t_size_op = custom_op(
     arg_types=[str_rprimitive],
     return_type=c_pyssize_t_rprimitive,
-    c_function_name='CPyStr_Size_size_t',
+    c_function_name='PyUnicode_GET_LENGTH',
     error_kind=ERR_NEG_INT)