--- a/mypyc/primitives/str_ops.py
+++ b/mypyc/primitives/str_ops.py
@@ -139,5 +139,5 @@ str_check_if_true = custom_op(
 str_ssize_t_size_op = custom_op(
     arg_types=[str_rprimitive],
     return_type=c_pyssize_t_rprimitive,
-    c_function_name='PyUnicode_GET_LENGTH',
+    c_function_name='CPyStr_Size_size_t',
     error_kind=ERR_NEG_INT)