--- a/mypyc/test-data/run-lists.test
+++ b/mypyc/test-data/run-lists.test
@@ -357,3 +357,7 @@ def test() -> None:
     source_e = [0, 1, 2]
     e = list((x ** 2) for x in (y + 2 for y in source_e))
     assert e == [4, 9, 16]
+    source_str = "abcd"
+    f = list("str:" + x for x in source_str)
+    assert f == ["str:a", "str:b", "str:c", "str:d"]
+