--- a/mypyc/test-data/run-tuples.test
+++ b/mypyc/test-data/run-tuples.test
@@ -196,9 +196,9 @@ def test_sequence_generator() -> None:
     a = tuple(f8(x) for x in source_fixed_length_tuple)
     assert a == (False, True, False, True)
 
-    source_str = "abbc"
-    a = tuple("s:" + x for x in source_str)
-    assert a == ("s:a", "s:b", "s:b", "s:c")
+    source_str = 'abbc'
+    b = tuple('s:' + x for x in source_str)
+    assert b == ('s:a', 's:b', 's:b', 's:c')
 
 TUPLE: Final[Tuple[str, ...]] = ('x', 'y')