### Description

Closes https://github.com/mypyc/mypyc/issues/835

* Add a branch for `str_rpimitive` in `builtin_len`
* Reduce redundant code
* Faster list/tuple built from str

## Test Plan

* Add run tests for faster list/tuple built from str
* Modify one irbuild test. The original one is redundant.