@JukkaL 2021-06-28
Can you double check the perf impact by running some quick microbenchmark?