@97littleleaf11 2021-06-28
This branch:
```
interpreted: 0.000373s (avg of 2601 iterations; stdev 1.6%)
compiled:    0.000180s (avg of 2601 iterations; stdev 1.3%)

compiled is 2.070x faster
```

Master branch:
```
interpreted: 0.000357s (avg of 2675 iterations; stdev 4.6%)
compiled:    0.000233s (avg of 2675 iterations; stdev 2.2%)

compiled is 1.529x faster
```

Tested on:
```py
def list_from_str() -> None:
    s1 = "aaaaa"
    n = 0
    for i in range(1000):
        list = [x for x in s1]
        n += len(list)
    assert n == 5000, n
```