@97littleleaf11 2021-06-29
Adding a wrapper function has little impact on performance, since the speed up is based on the optimized list comprehension:
```
interpreted: 0.000327s (avg of 2838 iterations; stdev 1.4%)
compiled:    0.000157s (avg of 2838 iterations; stdev 1.2%)

compiled is 2.084x faster
```